fix: [extension] circular dependency among context menu and utils (#3019)
* fix: [extension] circular dependency among contenxt menu and utils

* removed logs

* refactor: moved sendMessage away from utils

* remove unused types
