+++ browser-extension/mv3/src/service-worker/index.ts
@@ -3,7 +3,7 @@ import { registerCommands } from "./services/commands";
 import { initContextMenu } from "./services/contextMenu";
 import { initDevtoolsListener } from "./services/devtools";
 import { handleInstallUninstall } from "./services/installUninstall";
-import { initMessageHandler } from "./services/messageHandler";
+import { initMessageHandler } from "./services/messageHandler/listener";
 import { initRulesManager } from "./services/rulesManager";
 import { initWebRequestInterceptor } from "./services/webRequestInterceptor";
