+++ browser-extension/mv3/src/service-worker/services/contextMenu.ts
@@ -1,10 +1,6 @@
 import { onVariableChange, setVariable, Variable } from "../variable";
-import extensionIconManager from "./extensionIconManager";
 import { isExtensionEnabled } from "../../utils";
-
-// TODO: fix circular dependency
-import { sendMessageToApp } from "./messageHandler";
-import { CLIENT_MESSAGES } from "common/constants";
+import extensionIconManager from "./extensionIconManager";
 
 enum MenuItem {
   TOGGLE_ACTIVATION_STATUS = "toggle-activation-status",
@@ -25,8 +21,6 @@ export const updateActivationStatus = (isExtensionEnabled: boolean) => {
   } else {
     extensionIconManager.markExtensionDisabled();
   }
-
-  sendMessageToApp({ action: CLIENT_MESSAGES.NOTIFY_EXTENSION_STATUS_UPDATED, isExtensionEnabled });
 };
 
 export const initContextMenu = async () => {
