+++ browser-extension/mv3/src/service-worker/services/desktopApp/desktopAppManager.ts
@@ -1,11 +1,11 @@
 import { EXTENSION_MESSAGES } from "common/constants";
-import { sendMessageToApp } from "../messageHandler";
 import { applyProxy, ProxyDetails } from "../proxy";
 import { WebSocketManager } from "./webSocketManager";
 import { toggleExtensionStatus } from "../utils";
 import extensionIconManager from "../extensionIconManager";
 import { getConnectedBrowserAppId } from "./utils";
 import { DESKTOP_APP_CONFIG } from "./desktopAppConfig";
+import { sendMessageToApp } from "../messageHandler/sender";
 
 export class DesktopAppManager {
   private wsManager = new WebSocketManager();
