+++ browser-extension/mv3/src/service-worker/services/messageHandler/sender.ts
@@ -1,0 +1,6 @@
+import { tabService } from "../tabService";
+
+export const sendMessageToApp = async (messageObject: unknown) => {
+  const appTabs = await tabService.getAppTabs();
+  return Promise.all(appTabs.map(({ id }) => chrome.tabs.sendMessage(id, messageObject)));
+};
