+++ browser-extension/mv3/src/service-worker/services/rulesManager.ts
@@ -4,9 +4,9 @@ import { onVariableChange, Variable } from "../variable";
 import { debounce, getBlockedDomains, isExtensionEnabled, onBlockListChange } from "../../utils";
 import { TAB_SERVICE_DATA, tabService } from "./tabService";
 import { SessionRuleType } from "./requestProcessor/types";
-import { sendMessageToApp } from "./messageHandler";
 import { EXTENSION_MESSAGES } from "common/constants";
 import { UpdateDynamicRuleOptions } from "common/types";
+import { sendMessageToApp } from "./messageHandler/sender";
 
 // Object.values(chrome.declarativeNetRequest.ResourceType) cannot be used because in some browsers like safari and firefox
 // chrome.declarativeNetRequest.ResourceType is not defined.
